Refactor tokens main to split logic across multiple smaller components
