+++ .env.development
@@ -1,6 +1,6 @@
 #ADMIN
 NEXT_PUBLIC_CHAIN_ID=4
-NEXT_PUBLIC_RESERVOIR_API_BASE=https://api-rinkeby.reservoir.tools
+NEXT_PUBLIC_RESERVOIR_API_BASE=https://api.reservoir.tools
 NEXT_PUBLIC_PROXY_API_BASE=/api/reservoir
 #NEXT_PUBLIC_USE_WILDCARD=true
 #RESERVOIR_API_KEY={api-key}
