+++ components/ConnectWalletModal.tsx
@@ -25,7 +25,7 @@ const ConnectWalletModal: FC = () => {
       </Dialog.Trigger>
       <Dialog.Portal>
         <Dialog.Overlay>
-          <Dialog.Content className="fixed inset-0 bg-[#000000b6]">
+          <Dialog.Content className="fixed inset-0 z-[1000] bg-[#000000b6]">
             <div className="fixed top-1/2 left-1/2 w-[460px] -translate-x-1/2 -translate-y-1/2 transform rounded-2xl bg-white py-11 shadow-xl dark:bg-black">
               <div className="mx-12 mb-4 flex items-center justify-between">
                 <Dialog.Title className="reservoir-h4 font-headings dark:text-white">
