+++ components/InfoModal.tsx
@@ -13,7 +13,7 @@ const InfoModal: FC = () => {
       </Dialog.Trigger>
       <Dialog.Portal>
         <Dialog.Overlay>
-          <Dialog.Content className="fixed inset-0 bg-[#000000b6]">
+          <Dialog.Content className="fixed inset-0 bg-[#000000b6] z-[1000]">
             <div className="fixed top-1/2 left-1/2 max-w-prose -translate-x-1/2 -translate-y-1/2 transform rounded-md bg-white p-8 shadow-md">
               <div className="reservoir-body space-y-4">
                 <section>
